Fixed faulty book updating
Along with fixing the faulty book updating mechanics, I've also tested
the book generating feature. Because of the way I generate books,
instead of there being a 1 in 3 chance, it's always 1 in 1, because each
time a chunk is populated, the chests are rescanned and then most likely
populated with a book. To find naturally spawned chests easier, just
look in the console. You will see the coordinates of the chest if the
game creates one.